Adding a "Prefer Filename Date" option, and tweaking dates that Stash can't find
TPDB scrapers use the date of import for sites that don't have dates listed.  Obviously those can be much later than the actual release, however the releases tend to include a much closer date to the actual release date.  With this there's an optional config flag "prefer_file_date" (or -pfd on cli) to say "If the date in the filename is before the scraped date, then use the filename date instead"

Also a small date tweak to parse dates out of the filename that Stash doesn't pull correctly on its import